refactor : #44 이상징후 서비스를 전략 패턴 기반으로 점수 계산기를 호출하도록 수정
